requirements.txt
python-frontmatter
beautifulsoup4
markdownify
PyYAML